+++ include/multilevel_bisect/sample.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
-#include <vector>
 #include <random>
+#include <vector>
 
 #include "hypergraph/hypergraph.hpp"
 #include "options.hpp"
@@ -11,11 +11,12 @@ namespace pmondriaan {
 /**
  * Returns a vector of ns randomly selected sample vertices.
  */
-std::vector<int> sample_random(pmondriaan::hypergraph& H, int ns, std::mt19937& rng);
+std::vector<long> sample_random(pmondriaan::hypergraph& H, long ns, std::mt19937& rng);
 
 /**
  * Returns a vector of ns samples seleccted using the label propagation algorithm.
  */
-std::vector<int> sample_lp(pmondriaan::hypergraph& H, pmondriaan::options& opts, std::mt19937& rng);
+std::vector<long>
+sample_lp(pmondriaan::hypergraph& H, pmondriaan::options& opts, std::mt19937& rng);
 
 } // namespace pmondriaan
