+++ include/work_item.hpp
@@ -8,21 +8,21 @@ namespace pmondriaan {
  */
 class work_item {
   public:
-    work_item(int start, int end, int label_low, int label_high, long weight)
+    work_item(long start, long end, long label_low, long label_high, long weight)
     : start_(start), end_(end), label_low_(label_low), label_high_(label_high),
       weight_(weight) {}
 
-    int start() { return start_; }
-    int end() { return end_; }
-    int label_low() { return label_low_; }
-    int label_high() { return label_high_; }
+    long start() { return start_; }
+    long end() { return end_; }
+    long label_low() { return label_low_; }
+    long label_high() { return label_high_; }
     long weight() { return weight_; }
 
   private:
-    int start_;
-    int end_;
-    int label_low_;
-    int label_high_;
+    long start_;
+    long end_;
+    long label_low_;
+    long label_high_;
     long weight_;
 };
